libs/svt-av1/test/benchmarking/requirements.txt
pyyaml
pandas
matplotlib
meson
tqdm
numpy
scipy
